fix: config migration when --config-file flag is set

+++ packages/data-context/src/actions/MigrationActions.ts
@@ -9,20 +9,37 @@ import {
   SpecToMove,
   supportFilesForMigration,
 } from '../sources'
+import { getError } from '@packages/errors'
 
 export class MigrationActions {
   constructor (private ctx: DataContext) { }
 
   async createConfigFile () {
     const config = await this.ctx.migration.createConfigString()
+    const configFileNameAfterMigration = this.ctx.migration.configFileNameAfter
+
+    const checkIfFileExists = await this.ctx.actions.file.checkIfFileExists(configFileNameAfterMigration)
+
+    if (checkIfFileExists) {
+      this.ctx.onError(getError('LEGACY_CONFIG_FILE', configFileNameAfterMigration, this.ctx.lifecycleManager.projectRoot, this.ctx.lifecycleManager.legacyConfigFile))
+
+      return false
+    }
+
+    this.ctx.lifecycleManager.setConfigFilePath(configFileNameAfterMigration)
 
     await this.ctx.fs.writeFile(this.ctx.lifecycleManager.configFilePath, config).catch((error) => {
       throw error
     })
 
-    await this.ctx.actions.file.removeFileInProject('cypress.json').catch((error) => {
+    await this.ctx.actions.file.removeFileInProject(this.ctx.lifecycleManager.legacyConfigFile).catch((error) => {
       throw error
     })
+
+    // @ts-ignore configFile needs to be updated with the new one, so it finds the correct one
+    this.ctx.modeOptions.configFile = this.ctx.migration.configFileNameAfter
+
+    return true
   }
 
   initialize () {
